333 reload the url if the extension context gets invalidated when the extension is updated (#278)
* refactor: Simplify select option handling in WizardElementUtil

* Error handling on invalited

* feat: Add script to upload batch of files to Azure Storage

Added a new script `az-upload-batch.sh` to upload a batch of files to Azure Storage. This script uses the `az storage blob upload-batch` command to upload all files in the `./locales` directory to the `acf-i18n` container in the `` storage account. The `--auth-mode login` flag is used for authentication and the `--overwrite` flag is used to overwrite any existing files with the same name.

* refactor: Update French and Chinese translations

Update the French and Chinese translations in the `web-new.json` files. The changes include updating the translations for the languages "hi" and "zh-cn". This ensures that the translations are accurate and up to date.