329 target only the base url (#279)
* feat: Add URL match option in ConfigSettingsModal

The code changes add a new form group in the `ConfigSettingsModal` component to allow users to select the URL match option. This is achieved by adding a new property `url_match` to the `Configuration` type and updating the form group in the component. The commit also includes updates to the `URL_MATCH` enum in the `config-model.ts` file.

Recent user commits and repository commits are not relevant for generating the commit message.

* refactor: Remove unused files and configurations for acf-options-page-e2e

* refactor: Remove unused files and configurations for acf-options-page-e2e

* chore: Remove setupFiles from jest.config.ts