feat: Update NX_RELEASE_VERSION to 9.9.9 in acf-extension and acf-options-page
The code changes update the NX_RELEASE_VERSION environment variable to 9.9.9 in the .env.example files of the acf-extension and acf-options-page apps. This ensures that the correct release version is used in the applications. The webpack.config.js file in the acf-extension app is also updated to remove the 'v' prefix from the version number. Additionally, the environments.ts file in the acf-extension app is modified to remove the NX_RELEASE_VERSION export. These changes improve the consistency and accuracy of the release version across the project.