feat: Update NX_RELEASE_VERSION to 9.9.9 in acf-extension and acf-options-page (#280)
* feat: Update NX_RELEASE_VERSION to 9.9.9 in acf-extension and acf-options-page

The code changes update the NX_RELEASE_VERSION environment variable to 9.9.9 in the .env.example files of the acf-extension and acf-options-page apps. This ensures that the correct release version is used in the applications. The webpack.config.js file in the acf-extension app is also updated to remove the 'v' prefix from the version number. Additionally, the environments.ts file in the acf-extension app is modified to remove the NX_RELEASE_VERSION export. These changes improve the consistency and accuracy of the release version across the project.

* feat: Update NX_RELEASE_VERSION to 9.9.9 in acf-extension and acf-options-page

* Update permissions in deploy workflow

* removed provenance

* chore: Update NODE_AUTH_TOKEN in deploy workflow

The code changes update the NODE_AUTH_TOKEN environment variable in the deploy workflow. The previous value, ${{ secrets.GITHUB_TOKEN }}, is replaced with ${{ secrets.REPO_PAT }}. This change ensures that the correct authentication token is used for publishing packages during the deployment process.