332 i18n improve (#281)
* feat: Update language translations for Korean, Chinese, and French

The code changes update the language translations for Korean, Chinese, and French in the `web-new.json` files. The translations for the languages "hi" and "zh-cn" are modified to ensure accuracy and up-to-date translations. This improves the localization of the application for these languages.

* feat: Update context menu titles for field configuration, auto clicker, and configuration page

The code changes update the context menu titles for field configuration, auto clicker, and configuration page in the `registerContextMenus` function of the `context-menu.ts` file. The titles are now dynamically fetched from the i18n messages based on the user's language preference. This improves the localization of the context menu options for different languages.

* feat: Update context menu titles for field configuration, auto clicker, and configuration page

* feat: Update context menu titles for field configuration, auto clicker, and configuration page

* updated wizard

* updated env

* feat: Update context menu titles for field configuration, auto clicker, and configuration page

* moved backup

* Updated translate

* feat: Update context menu titles for field configuration, auto clicker, and configuration page

* feat: Update language translations for Korean, Chinese, and French

* chore(release): release $3.12.0

* Fixed formatting

* Fixed extension name

* Fixed format

---------

Signed-off-by: Dharmesh <dharmesh.hemaram@gmail.com>