Added license (#309)
* Added license

* Formated

* updated git tag version

* chore(release): release $4.0.9

* Formated

* updated web deploy command

* Added skip for version 4.0.9

* feat: Add conditional check for version before opening options tab

The code changes add a conditional check for the version before opening the options tab. This check ensures that the options tab is only opened if the version is not 4.0.9. This change improves the user experience by preventing the options tab from opening unnecessarily for a specific version.

Note: The recent user commits and recent repository commits do not provide any relevant information for crafting the commit message.

* Added skip for version 4.0.8